Reorganize code so that routing code is more encapsulated.
