+++ client/src/index.js
@@ -2,31 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import Home from './Home';
-import Banner from './Banner';
 import * as serviceWorker from './serviceWorker';
 import {
   BrowserRouter as Router,
-  Switch,
-  Route
 } from "react-router-dom";
 
 ReactDOM.render(
   <React.StrictMode>
-    <Banner />
-    <hr />
     <Router>
-      <Switch>
-
-        <Route exact path="/">
-          <Home />
-        </Route>
-
-        <Route path="/game">
-          <App />
-        </Route>
-
-      </Switch>
+      <App />
     </Router>
   </React.StrictMode>,
   document.getElementById('root')
